+++ query/result.go
@@ -1,6 +1,7 @@
 package query
 
 import (
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/arrow"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/result"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/scanner"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/types"
@@ -8,6 +9,7 @@ import (
 
 type (
 	Result            = result.Result
+	ArrowResult       = arrow.Result
 	ResultSet         = result.Set
 	ClosableResultSet = result.ClosableResultSet
 	Row               = result.Row